refactor: reorganize entry modals — remove aggregate, colocate with triggers, rename to _PascalCase
- New: _UploadEntryButton (button + New Entry modal + permission guard)
- Extracted: data-view/ (generic DataView + ListViewController)
- Moved: bulk modals inline to _EntryFilterToolbar with local state
- Renamed: 4 entry-only files to _PascalCase convention
- Deleted: entry-handle-modals.svelte aggregate
- Updated: +page.svelte (simplified), entry-dropdown-menu.svelte (import path)

Behavior-preserving per ClickUp 869duk832. All tooling clean (svelte-check, eslint, prettier).
