+++ app/frontend/src/lib/components/app/datasets/entries/overlays/_UploadEntryFormModal.svelte
